+++ addons/simple-mvp/src/main/java/org/vaadin/spring/mvp/Presenter.java
