+++ addons/simple-mvp/src/test/java/org/vaadin/spring/mvp/FooView.java
