+++ addons/simple-mvp/src/test/java/org/vaadin/spring/mvp/explicit/ExplicitPresenterIT.java
