+++ pom.xml
@@ -73,7 +73,7 @@
         <project.build.encoding>UTF-8</project.build.encoding>
         <project.build.target>1.6</project.build.target>
         <project.build.source>1.6</project.build.source>
-        <vaadin.version>7.4.0</vaadin.version>
+        <vaadin.version>7.4.1</vaadin.version>
         <vaadin-touchkit.version>4.0.0</vaadin-touchkit.version>
         <vaadin-spring.version>0.0.1-SNAPSHOT</vaadin-spring.version>
         <spring.version>4.1.4.RELEASE</spring.version>
@@ -306,34 +306,14 @@
             <id>vaadin-addons</id>
             <url>http://maven.vaadin.com/vaadin-addons</url>
         </repository>
-
-        <!-- Add Spring repositories -->
-        <!-- (you don't need this if you are using a .RELEASE version) -->
-        <!-- <repository> <id>spring-snapshots</id> <url>http://repo.spring.io/snapshot</url> 
-            <snapshots><enabled>true</enabled></snapshots> </repository> <repository> 
-            <id>spring-milestones</id> <url>http://repo.spring.io/milestone</url> </repository> -->
     </repositories>
 
-    <!-- <pluginRepositories> <pluginRepository> <id>spring-snapshots</id> 
-        <url>http://repo.spring.io/snapshot</url> </pluginRepository> <pluginRepository> 
-        <id>spring-milestones</id> <url>http://repo.spring.io/milestone</url> </pluginRepository> 
-        </pluginRepositories> -->
-
     <modules>
-<!--        <module>spring-boot-vaadin</module>
-        <module>spring-vaadin</module>
-        <module>spring-vaadin-security</module>
-        <module>spring-vaadin-touchkit</module>
-        <module>spring-vaadin-test</module>
-        <module>spring-vaadin-mvp</module>
-        <module>samples</module>
-        <module>spring-vaadin-simple-mvp</module>
-        <module>spring-vaadin-sidebar</module>
-        <module>spring-vaadin-i18n</module>-->
         <module>extensions/core</module>
         <module>extensions/test</module>
         <module>addons/eventbus</module>
         <module>addons/i18n</module>
         <module>addons/sidebar</module>
+        <module>addons/simple-mvp</module>
     </modules>
 </project>
