--- spring-vaadin-simple-mvp/src/test/java/org/vaadin/spring/mvp/autowired/ScanConfig.java
@@ -1,25 +1,0 @@
-/*
- * Copyright 2015 The original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.vaadin.spring.mvp.autowired;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.vaadin.spring.events.annotation.EnableVaadinEventBus;
-
-@ComponentScan
-@EnableVaadin
-@EnableVaadinEventBus
-public class ScanConfig {
-}
